Updated MTOW loop and module declaration
SizingMission is removed.
