+++ src/fastoad/register.py
@@ -24,13 +24,9 @@ from fastoad.models.handling_qualities.compute_static_margin import ComputeStati
 from fastoad.models.handling_qualities.tail_sizing.compute_tail_areas import ComputeTailAreas
 from fastoad.models.loops.compute_wing_area import ComputeWingArea
 from fastoad.models.performances.breguet import OMBreguet
+from fastoad.models.performances.mission.openmdao.link_mtow import LinkMTOW
 from fastoad.models.performances.mission.openmdao.mission import Mission
-from fastoad.models.performances.mission.openmdao.sizing_mission import SizingMission
-from fastoad.models.propulsion.fuel_propulsion.rubber_engine import (
-    OMRubberEngineComponent,
-    OMRubberEngineWrapper,
-)
-from fastoad.models.weight.mass_breakdown.mass_breakdown import MTOWComputation
+from fastoad.models.propulsion.fuel_propulsion.rubber_engine import OMRubberEngineComponent
 from fastoad.models.weight.weight import Weight
 from fastoad.module_management.constants import ModelDomain
 from fastoad.module_management.service_registry import RegisterPropulsion, RegisterOpenMDAOSystem
@@ -76,7 +72,9 @@ def register_openmdao_systems():
     # Loops #######################################################################
     RegisterOpenMDAOSystem("fastoad.loop.wing_area", domain=ModelDomain.OTHER)(ComputeWingArea)
 
-    RegisterOpenMDAOSystem("fastoad.loop.mtow", domain=ModelDomain.WEIGHT)(MTOWComputation)
+    RegisterOpenMDAOSystem(
+        "fastoad.mass_performances.link_MTOW", domain=ModelDomain.OTHER
+    )(LinkMTOW)
 
     # Weight ######################################################################
     RegisterOpenMDAOSystem("fastoad.weight.legacy", domain=ModelDomain.WEIGHT)(Weight)
@@ -85,13 +83,9 @@ def register_openmdao_systems():
         OMBreguet
     )
 
-    RegisterOpenMDAOSystem("fastoad.performances.sizing_mission", domain=ModelDomain.PERFORMANCE)(
-        SizingMission
-    )
-
-    OpenMDAOSystemRegistry.register_system(
-        Mission, "fastoad.performances.mission", domain=ModelDomain.PERFORMANCE
-    )
+    RegisterOpenMDAOSystem(
+         "fastoad.performances.mission", domain=ModelDomain.PERFORMANCE
+    )(Mission)
 
     # Propulsion ##################################################################
     rubber_engine_description = """
